requirements.txt
streamlit
opencv-python-headless
numpy
pandas
scipy
tqdm
plotly
PyYAML
torch
torchvision
timm
scikit-learn
joblib
Pillow
ultralytics
transformers
